+++ src/backup/sflash.c
@@ -231,7 +231,7 @@ sf_write_rom (const char *filename, unsigned short parport)
       return -1;
     }
 
-  size = ucon64.file_size;
+  size = (int) ucon64.file_size;
   printf ("Send: %d Bytes (%.4f Mb)\n\n", size, (float) size / MBIT);
 
   if (check_card () == 0)
@@ -364,7 +364,7 @@ sf_write_sram (const char *filename, unsigned short parport)
   void (*write_block) (int *, unsigned char *) = write_ram_by_byte; // write_ram_by_page
   (void) write_ram_by_page;
 
-  size = ucon64.file_size;
+  size = (int) ucon64.file_size;
   address = 0xfe0000;
 
   if ((file = fopen (filename, "rb")) == NULL)
