+++ src/console/nds.c
@@ -159,7 +159,7 @@ nds_n (st_ucon64_nfo_t *rominfo, const char *name)
   strncpy (buf, name, NDS_NAME_LEN);
   strcpy (dest_name, ucon64.fname);
   ucon64_file_handler (dest_name, NULL, 0);
-  fcopy (ucon64.fname, 0, ucon64.file_size, dest_name, "wb");
+  fcopy (ucon64.fname, 0, (size_t) ucon64.file_size, dest_name, "wb");
   ucon64_fwrite (buf, NDS_HEADER_START + rominfo->backup_header_len,
                  NDS_NAME_LEN, dest_name, "r+b");
 
@@ -175,7 +175,7 @@ nds_logo (st_ucon64_nfo_t *rominfo)
 
   strcpy (dest_name, ucon64.fname);
   ucon64_file_handler (dest_name, NULL, 0);
-  fcopy (ucon64.fname, 0, ucon64.file_size, dest_name, "wb");
+  fcopy (ucon64.fname, 0, (size_t) ucon64.file_size, dest_name, "wb");
   ucon64_fwrite (nds_logodata, NDS_HEADER_START + rominfo->backup_header_len +
                  192, NDS_LOGODATA_LEN, dest_name, "r+b");
 
@@ -192,7 +192,7 @@ nds_chk (st_ucon64_nfo_t *rominfo)
 
   strcpy (dest_name, ucon64.fname);
   ucon64_file_handler (dest_name, NULL, 0);
-  fcopy (ucon64.fname, 0, ucon64.file_size, dest_name, "wb");
+  fcopy (ucon64.fname, 0, (size_t) ucon64.file_size, dest_name, "wb");
 
   p = (unsigned char *) &rominfo->current_internal_crc;
   ucon64_fwrite (p, NDS_HEADER_START + rominfo->backup_header_len + 0x15e, 2,
